requirements.txt
langchain
langgraph
openai
chromadb
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
sentence-transformers
